+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.4
+
+* Update test
+  * Add deserializer context
+* Fix map subscriber error if field not exist in map array
+* Moving the entity link procedure
+* Fix if key is changed in collection relation
+
 # 0.1.3
 
 * Fix dependency injection for ``Coosos\VersionWorkflowBundle\EventListener\Doctrine\OnFlushListener``